pyproject.toml
[project]
name = "oryx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djoser>=2.3.1",
    "pillow>=11.2.1",
    "ruff>=0.12.0",
    "stripe>=12.2.0",
]
